[4 hours work BA]
-cardforge—improve card editing flow to be bug free.

bug1: clicking abilities area when abilities are empty should not bring up the buttons [fixed]
bug1.5: adding an ability should show the ability add/remove buttons [fixed]
bug2: should not allow user to click “ok” from abilities screen when the point cost exceeds allowed [fixed]
bug3 (minor): ability cost should not show up in red when it fits into costs on the main view—in progress
bug4—adding ability can sometimes reset cost to 0 on abilities edit screen, seems to not be refreshing correctly

this code appears to be involved in improperly extracting an ability cost of 0 when returning to the view.
AbilityWrapper*wrapper = self.currentAbilities[indexPath.row];

found this method for setting the ability points of a current ability wrapper.  Need to leverage this to reset the costs of abilities when returning to the abilityViewController

+(void)updateAbilityPoints:(CardModel*)card forWrapper:(AbilityWrapper*)wrapper withWrappers:(NSArray*)wrappers